Move DelimiterAwareTextSplitterTest to an algo module. Use it to also directly split the text from the reader items. This will probably cause some issues where users saved read positions of items that has text > 512 length. Will not do migration as it requires a large amount of testing to verify all edge cases of large text bodies. Will assume that current entries in DB were invalid, the reader will reset the state to a stable one after first read of the item.
