+++ build.gradle.kts
@@ -18,6 +18,7 @@ plugins {
     alias(libs.plugins.kotlin.serialization) apply false
     alias(libs.plugins.android.library) apply false
     alias(libs.plugins.dependency.analysis) apply false
+    alias(libs.plugins.jetbrains.kotlin.jvm) apply false
 }
 
 tasks.register("clean", Delete::class) {
