+++ settings.gradle.kts
@@ -52,3 +52,4 @@ include(":features:libraryExplorer")
 include(":tooling:application_workers")
 include(":features:webview")
 include(":tooling:local_source")
+include(":tooling:algorithms")
