+++ tooling/algorithms/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    alias(libs.plugins.noveldokusha.android.library)
+}
+
+android {
+    namespace = "my.noveldokusha.algorithms"
+}
+
+dependencies {
+    implementation(libs.test.junit)
+}
