+++ tooling/text_to_speech/build.gradle.kts
@@ -8,6 +8,8 @@ android {
 }
 
 dependencies {
+    implementation(projects.tooling.algorithms)
+
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
